Migrate to HAL library
Replaced files to match new library
Added UART-USB echo functionality
